+++ src/routes/router.tsx
@@ -1,51 +1,18 @@
-import { Flex } from '@/components/commons/Flex';
-import { SText } from '@/components/commons/SText';
-import { SimpleLoader } from '@/components/commons/SimpleLoader';
-import { Spacer } from '@/components/commons/Spacer';
 import { MultiSectionLayout } from '@/components/layout/MultiSectionHeader';
 import { SingleSectionLayout } from '@/components/layout/SingleSectionLayout';
 
-import { ComponentType, LazyExoticComponent, Suspense, lazy } from 'react';
+import { Suspense } from 'react';
 import { createBrowserRouter } from 'react-router-dom';
 
 import { Home } from '@/pages/Home/Home';
 import { TeamCalendarPage } from '@/pages/TeamCalendar/TeamCalendar';
 import { TeamDashboardPage } from '@/pages/TeamDashboard/TeamDashboard';
+import { LazyEnrollTemplate, LazySkeleton } from '@/routes/Lazies';
 import { PATH } from '@/routes/path';
 import { LoginTemplate } from '@/templates/Login/LoginTemplate';
 import { TeamModificationTemplate } from '@/templates/Team/TeamModificationTemplate';
 import { TeamRegistrationTemplate } from '@/templates/Team/TeamRegistrationTemplate';
 import { ModificationTemplate } from '@/templates/User/ModificationTemplate';
-
-const LazySkeleton = () => (
-  <Flex align="center" direction="column">
-    <Spacer h={150} />
-    <SimpleLoader />
-    <Spacer h={15} />
-    <SText>로딩중</SText>
-  </Flex>
-);
-
-/**
- * useState, useEffect 써서 MinTimePreservedFallback을 만드는 건 좀 별로 인거 같음.
- * 그냥 Promise.all을 사용하는게 그냥 로직상으로도, 코드상으로도 이해가 편할듯
- * lazy 로드 했는데 응답이 빨리 오면 화면이 깜빡거리는거 같은 상황 발생해서 fallback 최소 렌더링 시간 추가
- */
-const delayedLazy = (
-  importFunction: () => Promise<{ default: ComponentType<unknown> }>,
-  delay = 600
-): LazyExoticComponent<ComponentType<unknown>> => {
-  return lazy(() =>
-    Promise.all([
-      importFunction(),
-      new Promise((resolve) => setTimeout(resolve, delay)),
-    ]).then(([moduleExports]) => moduleExports)
-  );
-};
-
-const EnrollTemplate = delayedLazy(
-  () => import('@/templates/User/EnrollTemplate')
-);
 
 export const router = createBrowserRouter([
   {
@@ -81,7 +48,7 @@ export const router = createBrowserRouter([
         path: PATH.ENROLL,
         element: (
           <Suspense fallback={<LazySkeleton />}>
-            <EnrollTemplate />
+            <LazyEnrollTemplate />
           </Suspense>
         ),
       },
